refactor(hexagonal-cycle22): migrate LogicPort to crates/domain
- Add DomainError/DomainResult (thiserror) — pure, no mongodb import
- Add crates/domain/src/ports.rs with LogicPort (was LogicTrait)
- Re-export LogicPort from src/logic/traits.rs for BC
- crates/domain stays pure (0 mongodb imports)
