+++ crates/domain/Cargo.toml
@@ -8,3 +8,5 @@ serde = { version = "1.0.229", features = ["derive"] }
 serde_json = "1.0.151"
 chrono = { version = "0.4.38", features = ["serde"] }
 uuid = { version = "1.11.0", features = ["v4"] }
+async-trait = "0.1"
+thiserror = "1"
